% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_cols.R
\name{aggregate_cols}
\alias{aggregate_cols}
\title{Title}
\usage{
aggregate_cols(DT, functions, variables_function, variables_aggregation)
}
\arguments{
\item{DT}{}

\item{functions}{}

\item{variables_function}{}

\item{variables_aggregation}{}
}
\description{
Title
}
