% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colnames.R
\name{check_colnames}
\alias{check_colnames}
\title{Function checks if two tables has the same columns and returns any columns that are not present in one of the tables}
\usage{
check_colnames(DT1, DT2)
}
\arguments{
\item{DT1}{data.table; first table to inspect}

\item{DT2}{data.table; the second table to inspect}
}
\description{
Function checks if two tables has the same columns and returns any columns that are not present in one of the tables
}
