% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logger.R
\name{create_logger}
\alias{create_logger}
\title{Creates logger}
\usage{
create_logger(name, logging_level, log_file)
}
\arguments{
\item{name}{character; name of logger}

\item{logging_level}{numeric or character; see help to logging::setLevel}

\item{log_file}{character; path to file with logs}
}
