% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mysql_connection.R
\name{create_mysql_connection}
\alias{create_mysql_connection}
\title{Title}
\usage{
create_mysql_connection(host_name = config$host_name,
  db_name = config$db_name, creds_path = config$creds_path)
}
\arguments{
\item{host_name}{character; name of host to connect}

\item{db_name}{character; name of data base to connect;}

\item{creds_path}{character; path to credentials}
}
\value{
db_bi; mysql connection
}
\description{
Title
}
