% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_duplicates.R
\name{detect_duplicates}
\alias{detect_duplicates}
\title{Title}
\usage{
detect_duplicates(DT, cols)
}
\arguments{
\item{DT}{data.table; table to inspect}

\item{cols}{vector of characters; columns with primary key (for sure? :D)}
}
\description{
Title
}
