% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates_in_data_table_utilities.R
\name{divide_data_table_to_duplicated_and_not_duplicated}
\alias{divide_data_table_to_duplicated_and_not_duplicated}
\title{Title}
\usage{
divide_data_table_to_duplicated_and_not_duplicated(DT, cols)
}
\arguments{
\item{DT}{}

\item{cols}{}
}
\description{
Title
}
