% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_our_package.R
\name{install_our_package}
\alias{install_our_package}
\title{Title}
\usage{
install_our_package(creds_path = "~/pass/dcaplan_ab.txt", package_name,
  user = "analitycy_reklamy", domain = "gitlab.iiit.pl")
}
\arguments{
\item{creds_path}{}

\item{package_name}{}

\item{user}{}

\item{domain}{}
}
\description{
Title
}
