% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns_lower_case.R
\name{modify_columns_and_aggregate}
\alias{modify_columns_and_aggregate}
\title{Title}
\usage{
modify_columns_and_aggregate(DT, cols_to_modify, funs_chosen,
  aggregation_functions, aggregation_columns, aggregation_key)
}
\arguments{
\item{DT}{}

\item{cols_to_modify}{}

\item{aggregation_columns}{}

\item{aggregation_key}{}

\item{fun_chosen}{}
}
\description{
Title
}
