% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns_lower_case.R
\name{modify_columns_clean}
\alias{modify_columns_clean}
\title{Title}
\usage{
modify_columns_clean(DT, cols_to_modify, aggregation_functions,
  aggregation_columns, aggregation_key)
}
\arguments{
\item{DT}{data.table; table to aggregate}

\item{cols_to_modify}{vector of characters; see help to modify_columns_and_aggregate fun}

\item{aggregation_columns}{list of vectors of characters; see help to modify_columns_and_aggregate fun}

\item{aggregation_key}{vector of characters; see help to modify_columns_and_aggregate fun}
}
\description{
Title
}
