% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{produce_datatable}
\alias{produce_datatable}
\title{Title}
\usage{
produce_datatable(DT, cols_colour = NULL, cols_hidden = c(),
  cols_bold = c(), page_length = 100, row_call_back = NULL,
  footer_call_back = NULL, filter = list(position = "top", clear = T, plain
  = F), ...)
}
\arguments{
\item{DT}{data.table; table to show}

\item{cols_colour}{vector of characters; cols to colour}

\item{cols_hidden}{vector of characters; columns to hide}

\item{cols_bold}{vector of characters; cols to bold}

\item{page_length}{integer; how many rows to show}

\item{row_call_back}{character; js to pass to datatable}

\item{footer_call_back}{character; js to pass to datatable}

\item{filter}{character; where to place filter}

\item{...}{other arguments passed to datatable fun}
}
\description{
Title
}
