% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates_in_data_table_utilities.R
\name{return_all_duplicated_rows}
\alias{return_all_duplicated_rows}
\title{Title}
\usage{
return_all_duplicated_rows(DT, cols)
}
\arguments{
\item{DT}{data.table; table to inspect}

\item{cols}{vector of characters; columns with primary key}
}
\description{
Title
}
