% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_if_duplicated_cols.R
\name{stop_if_duplicated_cols}
\alias{stop_if_duplicated_cols}
\title{Title}
\usage{
stop_if_duplicated_cols(DT)
}
\arguments{
\item{DT}{data.table; table for check if there are duplicated cols present}
}
\description{
Title
}
