% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_several_times.R
\name{try_several_times}
\alias{try_several_times}
\title{Title}
\usage{
try_several_times(fun, args, condition, tries = 15, wait = 5)
}
\arguments{
\item{fun}{function; fun to apply several times}

\item{args}{list of named characters; arguments of fun}

\item{condition}{expression; e.g. expression(out$status_code == 200)}

\item{tries}{integer; number of tries}

\item{wait}{integer; time to wait between tries in seconds}
}
\description{
Title
}
