% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_cols.R
\name{aggregate_cols}
\alias{aggregate_cols}
\title{Title}
\usage{
aggregate_cols(DT, functions, variables_function, variables_aggregation)
}
\arguments{
\item{DT}{data.table; table to aggregate}

\item{functions}{list of functions; functions to apply to DT}

\item{variables_function}{list of characters; variables functions would be applied to, accordingly (1st function to the first element of list, etc)}

\item{variables_aggregation}{vector of characters; variables for aggregation}
}
\value{
DT_out data.table; aggregated table
}
\description{
Title
}
