% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_our_package.R
\name{install_our_package}
\alias{install_our_package}
\title{Title}
\usage{
install_our_package(creds_path = "~/pass/git.txt", package_name,
  user = "analitycy_reklamy", domain)
}
\arguments{
\item{creds_path}{character; path to credentials}

\item{package_name}{character; package name to install}

\item{user}{character; gitlab user}

\item{domain}{character; gitlab_domain}
}
\description{
Title
}
