% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns_lower_case.R
\name{modify_columns_and_aggregate}
\alias{modify_columns_and_aggregate}
\title{Title}
\usage{
modify_columns_and_aggregate(DT, cols_to_modify, funs_chosen,
  aggregation_functions, aggregation_columns, aggregation_key)
}
\arguments{
\item{DT}{data.table; table to aggregate}

\item{cols_to_modify}{vector of characters; columns to modify by funs_chosen before aggregation}

\item{funs_chosen}{list of functions; functions applied to data before aggregate}

\item{aggregation_functions}{list of functions; functions to apply to aggregation_columns, accordingly}

\item{aggregation_columns}{list of vectors of characters; columns to aggregate}

\item{aggregation_key}{vector of characters; columns to aggregate by}
}
\description{
Title
}
