% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_columns_lower_case.R
\name{modify_columns_clean}
\alias{modify_columns_clean}
\title{Title}
\usage{
modify_columns_clean(DT, cols_to_modify, aggregation_functions,
  aggregation_columns, aggregation_key)
}
\arguments{
\item{DT}{}

\item{cols_to_modify}{}

\item{aggregation_functions}{}

\item{aggregation_columns}{}

\item{aggregation_key}{}
}
\description{
Title
}
