% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{produce_datatable}
\alias{produce_datatable}
\title{Title}
\usage{
produce_datatable(DT, cols_colour = NULL, cols_hidden = c(),
  cols_bold = c(), page_length = 100, row_call_back = NULL,
  footer_call_back = NULL, filter = list(position = "top", clear = T, plain
  = F), ...)
}
\arguments{
\item{DT}{}

\item{cols_colour}{}

\item{cols_hidden}{}

\item{cols_bold}{}

\item{page_length}{}

\item{row_call_back}{}
}
\description{
Title
}
