% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_several_times.R
\name{try_several_times}
\alias{try_several_times}
\title{Title}
\usage{
try_several_times(fun, args, condition, tries = 15, wait = 5)
}
\arguments{
\item{fun}{}

\item{args}{}

\item{condition}{}

\item{tries}{}

\item{wait}{}
}
\description{
Title
}
